Factorized projectilesystem because it looked bad. Added some functionality for enemy json projectiles. All projectiles now have hover component and should not miss their target.

+++ src/world/component/guncomponent.hpp
@@ -11,8 +11,8 @@ struct GunComponent : public Component {
 	virtual ~GunComponent();
 
 	GunType type;
-	int cooldown;
-	int cooldownLength;
+	float cooldown;
+	float cooldownLength;
 	bool shoot;
 
 	void addGun(GunType inType, int inCooldown);
